backend/requirements.txt
Django==2.2.5
djangorestframework==3.10.3
pytz==2019.3
sqlparse==0.3.0
